fix: distance type changed to Float, Double is overkill feat: exception throwing on insert route validation feat: more on user history, fails in getting the shortest path

+++ src/main/java/cujae/edmaps/core/dijkstra/DijkstraShortestPath.java
@@ -12,13 +12,13 @@ public class DijkstraShortestPath {
     private Vertex start;
     private Map<Vertex, WayToArrive> nodes;
 
-    private class WayToArrive{
+    private class WayToArrive {
         private Vertex previous;
-        private Double distance;
+        private Float distance;
 
         private Bus bus;
 
-        public WayToArrive(Vertex previous, Double distance, Bus bus) {
+        public WayToArrive(Vertex previous, Float distance, Bus bus) {
             this.previous = previous;
             this.distance = distance;
             this.bus = bus;
@@ -29,18 +29,18 @@ public class DijkstraShortestPath {
         this.start = start;
         nodes = new HashMap<>();
         Map<Vertex, WayToArrive> toUnlock = new HashMap<>();
-        toUnlock.put(start, new WayToArrive(null, 0.0, null));
+        toUnlock.put(start, new WayToArrive(null, 0.0f, null));
         dijkstraAlgorithm(new HashSet<>(), toUnlock);
     }
 
-    public void dijkstraAlgorithm(Set<Vertex> alreadyUnlocked, Map<Vertex, WayToArrive> toUnlock){
+    public void dijkstraAlgorithm(Set<Vertex> alreadyUnlocked, Map<Vertex, WayToArrive> toUnlock) {
         Vertex vertex = getShortest(toUnlock);
-        while(vertex != null) {
-            Double distance = nodes.get(vertex).distance;
+        while (vertex != null) {
+            Float distance = nodes.get(vertex).distance;
             for (Edge e : vertex.getEdgeList()) {
                 Route weight = (Route) ((WeightedEdge) e).getWeight();
                 Vertex target = e.getVertex();
-                Double actualDistance = distance + weight.getDistance();
+                Float actualDistance = distance + weight.getDistance();
                 WayToArrive wayToArrive = null;
                 if (nodes.containsKey(target)) {
                     wayToArrive = nodes.get(target);
@@ -64,29 +64,29 @@ public class DijkstraShortestPath {
     }
 
     public CompletePath getShortestPathTo(Vertex goal) throws Exception {
-        CompletePath result =  new CompletePath();
-        while(goal != null){
-            if(nodes.containsKey(goal)){
+        CompletePath result = new CompletePath();
+        while (goal != null) {
+            if (nodes.containsKey(goal)) {
                 WayToArrive wayToArrive = nodes.get(goal);
                 result.addPath(goal, wayToArrive.bus, wayToArrive.distance);
                 goal = wayToArrive.previous;
-            }
-            else throw new Exception();
+            } else throw new Exception();
         }
         return result;
     }
-    private Vertex getShortest(Map<Vertex, WayToArrive> map){
+
+    private Vertex getShortest(Map<Vertex, WayToArrive> map) {
         Vertex result = null;
         Iterator<Map.Entry<Vertex, WayToArrive>> it = map.entrySet().iterator();
-        Double min = 0.0;
-        if(it.hasNext()){
+        Float min = 0.0f;
+        if (it.hasNext()) {
             Map.Entry<Vertex, WayToArrive> entry = it.next();
             min = entry.getValue().distance;
             result = entry.getKey();
         }
-        while(it.hasNext()){
+        while (it.hasNext()) {
             Map.Entry<Vertex, WayToArrive> entry = it.next();
-            if(min > entry.getValue().distance){
+            if (min > entry.getValue().distance) {
                 min = entry.getValue().distance;
                 result = entry.getKey();
             }
